+++ osu.Game.Tournament/Screens/Setup/StablePathSelectScreen.cs
@@ -35,7 +35,7 @@ namespace osu.Game.Tournament.Screens.Setup
         [BackgroundDependencyLoader(true)]
         private void load(Storage storage, OsuColour colours)
         {
-            var initialStorage = (ipc as FileBasedIPC)?.IPCStorage ?? storage;
+            var initialStorage = (ipc as FileAndGosuBasedIPC)?.IPCStorage ?? storage;
             string initialPath = new DirectoryInfo(initialStorage.GetFullPath(string.Empty)).Parent?.FullName;
 
             AddRangeInternal(new Drawable[]
@@ -129,7 +129,7 @@ namespace osu.Game.Tournament.Screens.Setup
         protected virtual void ChangePath()
         {
             string target = directorySelector.CurrentPath.Value.FullName;
-            var fileBasedIpc = ipc as FileBasedIPC;
+            var fileBasedIpc = ipc as FileAndGosuBasedIPC;
             Logger.Log($"Changing Stable CE location to {target}");
 
             if (!fileBasedIpc?.SetIPCLocation(target) ?? true)
@@ -146,7 +146,7 @@ namespace osu.Game.Tournament.Screens.Setup
 
         protected virtual void AutoDetect()
         {
-            var fileBasedIpc = ipc as FileBasedIPC;
+            var fileBasedIpc = ipc as FileAndGosuBasedIPC;
 
             if (!fileBasedIpc?.AutoDetectIPCLocation() ?? true)
             {
